- player fixed. Death UI

+++ LudumDare2020/Assets/Scripts/PlayerHUD.cs
@@ -65,7 +65,7 @@ public class PlayerHUD : MonoBehaviour
 
             if(RestartIn)
             {
-                RestartIn.SetText("Restarting in: " + RestartTimer.ToString());
+                RestartIn.SetText("Restarting in: " + ((int)RestartTimer).ToString());
             }
         }
